requirements.txt
playwright
supabase
python-dotenv
requests
openai>=1.0.0
boto3